cleanign

+++ application/controllers/manage/Attributepolicy.php
@@ -146,6 +146,7 @@ class Attributepolicy extends MY_Controller
                 }
             }
         }
+
         return $this->output->set_content_type('application/json')->set_output(json_encode($result));
     }
 
@@ -297,26 +298,27 @@ class Attributepolicy extends MY_Controller
         } catch (Exception $e) {
             return $this->output->set_status_header(403)->set_output($e->getMessage());
         }
-
-        $isLocked = $ent->getLocked();
-        if ($isLocked) {
+        if ($ent->getLocked()) {
             return $this->output->set_status_header(403)->set_output('Entity is locked');
         }
         if ($this->updateattrspValidate() !== true) {
             return $this->output->set_status_header(403)->set_output('Incorrect data input');
         }
-        $attrid = trim($this->input->post('attrid'));
-        $policy = trim($this->input->post('policy'));
-        $spid = trim($this->input->post('spid'));
-        $customenable = trim($this->input->post('customenabled'));
-        $custompolicy = trim($this->input->post('custompolicy'));
-        $customvals = trim($this->input->post('customvals'));
+        $attrid = $this->input->post('attrid');
+        $policy = $this->input->post('policy');
+        $spid = $this->input->post('spid');
+        $customenable = $this->input->post('customenabled');
+        $custompolicy = $this->input->post('custompolicy');
+        $customvals = $this->input->post('customvals');
 
 
-        if (!ctype_digit($spid) || !ctype_digit($attrid) || !ctype_digit($policy) || !in_array($policy, array('0', '1', '2', '100'))) {
+        if (!in_array($policy, array('0', '1', '2', '100'))) {
             return $this->output->set_status_header(403)->set_output('Posted invalid data');
         }
         try {
+            /**
+             * @var models\AttributeReleasePolicy $controlCheck
+             */
             $controlCheck = $this->em->getRepository('models\AttributeReleasePolicy')->findOneBy(array('idp' => $ent->getId(), 'attribute' => $attrid, 'requester' => $spid, 'type' => array('sp', 'customsp')));
         } catch (Exception $e) {
             log_message('error', __METHOD__ . ' ' . $e);
@@ -333,14 +335,10 @@ class Attributepolicy extends MY_Controller
          * @var $sp models\Provider
          */
         $attribute = $this->em->getRepository('models\Attribute')->findOneBy(array('id' => $attrid));
-        $sp = $this->em->getRepository('models\Provider')->findOneBy(array('id' => $spid));
+        $sp = $this->em->getRepository('models\Provider')->findOneBy(array('id' => $spid, 'type' => array('SP', 'BOTH')));
 
-        if ($attribute === null) {
-            return $this->output->set_status_header(403)->set_output('Attribute passed in post does not exist');
-        }
-        $entype = $sp->getType();
-        if ($sp === null || !in_array($entype, array('SP', 'BOTH'))) {
-            return $this->output->set_status_header(403)->set_output('SP passed in post does not exist or not valid entity type');
+        if ($attribute === null || $sp === null) {
+            return $this->output->set_status_header(403)->set_output('Attribute or SP passed in post does not exist');
         }
 
         if ($policy !== '100') {
@@ -348,7 +346,7 @@ class Attributepolicy extends MY_Controller
             $attrPolicy->setSpecificPolicy($ent, $attribute, $spid, $policy);
             $this->em->persist($attrPolicy);
         }
-        if (!empty($customenable) && $customenable === 'yes' && !empty($custompolicy) && in_array($custompolicy, array('permit', 'deny')) && !empty($customvals)) {
+        if ($customenable === 'yes' && !empty($custompolicy) && in_array($custompolicy, array('permit', 'deny'), true) && !empty($customvals)) {
             $customRawDataArray = array();
             $explcustomvals = explode(',', $customvals);
             foreach ($explcustomvals as $r) {
@@ -372,15 +370,13 @@ class Attributepolicy extends MY_Controller
 
         try {
             $this->em->flush();
-
-            return $this->output->set_content_type('application/json')->set_output(json_encode(array('status' => 'success')));
-
         } catch (Exception $e) {
             log_message('error', __METHOD__ . ' ' . $e);
 
             return $this->output->set_status_header(500)->set_output('Internal Server Error');
         }
 
+        return $this->output->set_content_type('application/json')->set_output(json_encode(array('status' => 'success')));
 
     }
 
@@ -672,8 +668,7 @@ class Attributepolicy extends MY_Controller
             return $this->output->set_status_header(403)->set_output($e->getMessage());
         }
 
-        $isLocked = $ent->getLocked();
-        if ($isLocked) {
+        if ($ent->getLocked()) {
             return $this->output->set_status_header(403)->set_output('Entity is locked');
         }
         $attrid = trim($this->input->post('attrid'));
@@ -698,10 +693,8 @@ class Attributepolicy extends MY_Controller
          * @var $attrPolicy models\AttributeReleasePolicy
          */
         $attrPolicy = $this->em->getRepository('models\AttributeReleasePolicy')->findOneBy(array('attribute' => $attrid, 'idp' => $ent->getId(), 'type' => 'fed', 'requester' => $fedid));
-        if ($policy === '100') {
-            if ($attrPolicy !== null) {
-                $this->em->remove($attrPolicy);
-            }
+        if ($policy === '100' && $attrPolicy !== null) {
+            $this->em->remove($attrPolicy);
         } elseif ($attrPolicy === null) {
             $attrPolicy = new models\AttributeReleasePolicy();
             if ($federation === null) {
@@ -714,16 +707,15 @@ class Attributepolicy extends MY_Controller
             $this->em->persist($attrPolicy);
         }
 
-
         try {
             $this->em->flush();
-
-            return $this->output->set_content_type('application/json')->set_output(json_encode(array('status' => 'success')));
         } catch (Exception $e) {
             log_message('error', __METHOD__ . ' ' . $e);
 
             return $this->output->set_status_header(500)->set_output('Internal Server Error');
         }
+
+        return $this->output->set_content_type('application/json')->set_output(json_encode(array('status' => 'success')));
 
     }
 
@@ -732,8 +724,10 @@ class Attributepolicy extends MY_Controller
         $this->form_validation->set_rules('customvals', '' . lang('permdenvalue') . '', 'trim|alpha_dash_comma');
         $this->form_validation->set_rules('custompolicy', 'Custom Policy', 'trim');
         $this->form_validation->set_rules('customenabled', 'Custom enabled', 'trim');
-        $this->form_validation->set_rules('attrid', 'Attribute', 'trim|required|numeric');
-        $this->form_validation->set_rules('spid', 'Service Provider ID', 'trim|required|numeric');
+        $this->form_validation->set_rules('attrid', 'Attribute', 'trim|required|is_natural_no_zero');
+        $this->form_validation->set_rules('spid', 'Service Provider ID', 'trim|required|is_natural_no_zero');
+        $this->form_validation->set_rules('policy', 'Policy', 'trim|is_natural');
+
 
         return $this->form_validation->run();
     }
@@ -796,10 +790,8 @@ class Attributepolicy extends MY_Controller
             $this->em->persist($attrPolicy);
         }
 
-        if (empty($customenabled) || $customenabled !== 'yes' || empty($customvals)) {
-            if ($customattrPolicy !== null) {
-                $this->em->remove($customattrPolicy);
-            }
+        if (($customattrPolicy !== null) && (empty($customenabled) || $customenabled !== 'yes' || empty($customvals))) {
+            $this->em->remove($customattrPolicy);
         } else {
             $valsarray = array();
             $cvalsExploded = explode(',', $customvals);
